refactor: move keyboard event handling to Search component for better encapsulation
